# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/posts":{"post":{"operationId":"PostsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDto"}}}},"responses":{"201":{"description":"게시글 생성 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}},"401":{"description":"인증되지 않은 사용자"}},"security":[{"Authorization":[]}],"summary":"게시글 생성","tags":["api/v1/posts"]},"get":{"operationId":"PostsController_findAll","parameters":[{"name":"theme","required":false,"in":"query","description":"필터링할 게시글 테마","schema":{"type":"string"}},{"name":"category","required":false,"in":"query","description":"필터링할 게시글 카테고리","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","description":"페이지 번호 (기본값: 1)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"페이지당 게시글 수 (기본값: 10)","schema":{"type":"string"}}],"responses":{"200":{"description":"게시글 목록 조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResponseDto"}}}}}},"summary":"게시글 목록 조회","tags":["api/v1/posts"]}},"/api/v1/posts/{id}":{"get":{"operationId":"PostsController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"MongoDB ObjectId (24자리 16진수 문자열, 예: 507f1f77bcf86cd799439011)","schema":{"example":"507f1f77bcf86cd799439011","type":"string"}},{"name":"x-viewed","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"게시글 상세 조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"summary":"게시글 상세 조회","tags":["api/v1/posts"]},"patch":{"operationId":"PostsController_update","parameters":[{"name":"id","required":true,"in":"path","description":"MongoDB ObjectId","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDto"}}}},"responses":{"200":{"description":"게시글 수정 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}},"401":{"description":"인증되지 않은 사용자"},"403":{"description":"수정 권한이 없는 사용자"},"404":{"description":"게시글을 찾을 수 없음"}},"security":[{"Authorization":[]}],"summary":"게시글 수정","tags":["api/v1/posts"]},"delete":{"operationId":"PostsController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"MongoDB ObjectId","schema":{"type":"string"}}],"responses":{"204":{"description":"게시글 삭제 성공"},"401":{"description":"인증되지 않은 사용자"}},"security":[{"Authorization":[]}],"summary":"게시글 삭제","tags":["api/v1/posts"]}},"/api/v1/posts/{id}/edit":{"get":{"operationId":"PostsController_findForEdit","parameters":[{"name":"id","required":true,"in":"path","description":"MongoDB ObjectId (24자리 16진수 문자열, 예: 507f1f77bcf86cd799439011)","schema":{"example":"507f1f77bcf86cd799439011","type":"string"}}],"responses":{"200":{"description":"게시글 수정용 조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}},"401":{"description":"인증되지 않은 사용자"},"403":{"description":"작성자가 아닌 사용자"},"404":{"description":"게시글을 찾을 수 없음"}},"security":[{"Authorization":[]}],"summary":"게시글 수정용 조회","tags":["api/v1/posts"]}}},"info":{"title":"VansDevBlog API","description":"VansDevBlog Post Service API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateDto":{"type":"object","properties":{"title":{"type":"string","example":"게시글 제목","description":"게시글의 제목"},"content":{"type":"object","example":{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"게시글 내용입니다."}]}]},"description":"게시글의 본문 내용 (Object)"},"theme":{"type":"string","example":"nest.js","description":"게시글의 테마"},"description":{"type":"string","example":"게시글 설명입니다.","description":"게시글의 설명"},"tags":{"example":["태그1","태그2"],"description":"게시글의 태그 목록","type":"array","items":{"type":"string"}},"category":{"type":"string","example":"introduction","description":"게시글의 카테고리"},"topic":{"type":"string","example":"Java 알고리즘","description":"게시글의 주제"},"thumbnail":{"type":"string","example":"thumbnail.jpg","description":"게시글의 썸네일 이미지"},"language":{"type":"string","example":"ko","description":"게시글의 언어"}},"required":["title","content","theme","description","tags","category","topic","thumbnail","language"]},"ObjectId":{"type":"object","properties":{}},"ResponseDto":{"type":"object","properties":{"_id":{"example":"507f1f77bcf86cd799439011","description":"MongoDB ObjectId","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"title":{"type":"string","example":"게시글 제목"},"content":{"type":"object","example":{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"게시글 내용입니다."}]}]},"description":"게시글 내용 (Object)"},"theme":{"type":"string","example":"nestjs"},"authorEmail":{"type":"string","example":"user@example.com"},"author":{"type":"string","example":"닉네임"},"createdAt":{"type":"string","example":"2025. 3. 24. 오후 9:14:10"},"updatedAt":{"type":"string","example":"2025. 4. 27. 오후 7:47:51"},"description":{"type":"string","example":"게시글 설명입니다."},"tags":{"example":["태그1","태그2"],"type":"array","items":{"type":"string"}},"viewCount":{"type":"number","example":0},"likeCount":{"type":"number","example":0},"category":{"type":"string","example":"introduction"},"thumbnail":{"type":"string","example":"thumbnail.jpg"},"language":{"type":"string","example":"ko"},"topic":{"type":"string","example":"Java 알고리즘"}},"required":["_id","title","content","theme","authorEmail","author","createdAt","updatedAt","description","tags","viewCount","likeCount","category","thumbnail","language","topic"]},"UpdateDto":{"type":"object","properties":{"title":{"type":"string","example":"수정된 제목","description":"수정할 게시글 제목"},"content":{"type":"object","example":{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"수정된 내용입니다."}]}]},"description":"수정할 게시글 내용 (Object)"},"theme":{"type":"string","example":"light","description":"수정할 테마"},"description":{"type":"string","example":"수정된 설명입니다.","description":"수정할 게시글 설명"},"tags":{"example":["태그1","태그2"],"description":"수정할 태그 목록","type":"array","items":{"type":"string"}},"category":{"type":"string","example":"introduction","description":"수정할 카테고리"},"topic":{"type":"string","example":"Java 알고리즘","description":"수정할 주제"},"thumbnail":{"type":"string","example":"thumbnail.jpg","description":"수정할 썸네일 이미지"},"language":{"type":"string","example":"ko","description":"수정할 언어"}}}}}}